update ProductDto to use BigDecimal for productPrice

+++ src/main/java/models/ProductDto.java
@@ -1,4 +1,6 @@
 package models;
 
-public record ProductDto(String productName, double productPrice) {
+import java.math.BigDecimal;
+
+public record ProductDto(String productName, BigDecimal productPrice) {
 }
